Get rid of the dependency on UTCTime.
Previously, we had to do through a lot of shenanigans to keep this code pure,
because traditionally LRUCaches record date time stamps for read/write times, so
you know what value is the least-recently-used.

Well, timestamps are a convenient way to do this, but really, all that matters
is the ordering. Numbers are a great way to keep track of relative ordering,
just make the access time a monotonically increasing number.

There's a risk of hitting integer overflow if this cache lives long enough, but
that can be fixed by either periodically subtracting the smallest access time
from all access times, or just setting the data type to Integer.